Miglioramento sensori e invio dati

+++ client/Client/Models/Detection.cs
@@ -9,7 +9,7 @@ namespace Client.Models
     {
         public int Speed { get; set; }
         public int Battery { get; set; }
-        public decimal Longitude { get; set; }
-        public decimal Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
     }
 }
